refactor: relocate schemas and remove cors
