+++ functions/index.js
@@ -1,9 +1,8 @@
 const functions = require('firebase-functions');
 const admin = require('firebase-admin');
 
-const cors = require('cors')({ origin: true });
 const orderLifecycle = require('./orderLifecycleService');
-const { Schemas, validateData } = require('../src/lib/models/schemas.js');
+const { Schemas, validateData } = require('./schemas.js');
 
 // Initialize Firebase Admin
 admin.initializeApp();
