+++ functions/package.json
@@ -17,8 +17,7 @@
     "firebase-admin": "^12.0.0",
     "firebase-functions": "^4.8.0",
     "stripe": "^14.0.0",
-    "@paypal/checkout-server-sdk": "^1.0.3",
-    "cors": "^2.8.5"
+    "@paypal/checkout-server-sdk": "^1.0.3"
   },
   "devDependencies": {
     "eslint": "^8.15.0",
