+++ src/lib/services/CheckoutService.js
@@ -6,13 +6,15 @@ import { Order } from '../models/Order.js';
 import { OrderItem } from '../models/OrderItem.js';
 import { Payment } from '../models/Payment.js';
 import { Shipping } from '../models/Shipping.js';
-import { Schemas, validateData } from '../models/schemas.js';
+import schemas from '../../../functions/schemas.js';
 import { errorHandler } from '../errorHandler.js';
 import OrderService from './OrderService.js';
 import PaymentService from './PaymentService.js';
 import ShippingService from './ShippingService.js';
 import ProductService from './ProductService.js';
 import logger from '../logger.js';
+
+const { Schemas, validateData } = schemas;
 
 export class CheckoutService {
   constructor() {
